requirements.txt
Flask==2.1.1
Werkzeug==2.1.1
psycopg2-binary==2.9.1
Flask-Caching==1.10.1
prometheus-flask-exporter==0.20.1